feat: move add dataset example to be a part of the examples tab
